feat: split home page into sub components
