--- src/i18n/home/home_en.json
@@ -1,31 +1,0 @@
-{
-    "dashboard": {
-        "dataset": "Dataset",
-        "service": "Service",
-        "dataspace": {
-            "summary": {
-                "counters": {
-                    "members": "Members",
-                    "see": "See",
-                    "rulebook":"The rulebook"
-                }
-            },
-            "creation": "Create a new Data Space",
-            "content": {
-                "new": "New"
-            },
-            "options": {
-                "add": "Add",
-                "new-dataset": "New dataset",
-                "new-dashboard": "New dashboard",
-                "new-service": "New service",
-                "new-report": "New report",
-                "visualize": "Visualize",
-                "datasets": "Datasets",
-                "services": "Services",
-                "dashboards":"Dashboards",
-                "reports":"Reports"
-            }
-        }
-    }
-}
