+++ src/i18n/index.ts
@@ -2,14 +2,14 @@ import { loadTranslations } from '@okp4/ui'
 
 import footer_en from './footer/footer_en.json'
 import footer_fr from './footer/footer_fr.json'
-import home_en from './home/home_en.json'
-import home_fr from './home/home_fr.json'
+import dashboard_en from './dashboard/dashboard_en.json'
+import dashboard_fr from './dashboard/dashboard_fr.json'
 
 const translationsToLoad = [
   { lng: 'en', namespace: 'footer', resource: footer_en },
   { lng: 'fr', namespace: 'footer', resource: footer_fr },
-  { lng: 'en', namespace: 'home', resource: home_en },
-  { lng: 'fr', namespace: 'home', resource: home_fr }
+  { lng: 'en', namespace: 'dashboard', resource: dashboard_en },
+  { lng: 'fr', namespace: 'dashboard', resource: dashboard_fr }
 ]
 
 loadTranslations(translationsToLoad)
